+++ CHANGELOG.rst
@@ -58,6 +58,8 @@ in development
 * The ``--tasks`` option in the CLI for ``st2 execution get`` and ``st2 run`` will be renamed to
   ``--show-tasks`` to avoid conflict with the tasks option in st2 execution re-run.
 * Add option to rerun one or more tasks in mistral workflow that has errored. (new-feature)
+* Fix a bug when removing notify section from an action meta and registering it never removed
+  the notify section from the db. (bug fix)
 * Make sure action specific short lived authentication token is deleted immediately when execution
   is canceled. (improvement)